* commit 'd298cba513a95603f8a898a3e9ed717bd436727e': (26 commits)
  Merge pull request #193 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0 to release/2.7.0
  Merge pull request #192 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0 to release/2.7.0
  Merge pull request #191 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0 to release/2.7.0
  Merge pull request #190 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0-rename to release/2.7.0
  Merge pull request #189 in ADUC/open-apaas-extapp-ios from feature/2.7.0_publish to release/2.7.0
  Merge pull request #188 in ADUC/open-apaas-extapp-ios from feature/2.7.0_scene_name to release/2.7.0
  Merge pull request #187 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0-ui to release/2.7.0
  Merge pull request #186 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0-loading to release/2.7.0
  Merge pull request #185 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0_loading to release/2.7.0
  Merge pull request #184 in ADUC/open-apaas-extapp-ios from feature/2.7.0-ui to release/2.7.0
  Merge pull request #183 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0-red to release/2.7.0
  Merge pull request #182 in ADUC/open-apaas-extapp-ios from feature/2.7.0-media to release/2.7.0
  Merge pull request #181 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0-fix to release/2.7.0
  Merge pull request #179 in ADUC/open-apaas-extapp-ios from feature/counterFix to release/2.7.0
  Merge pull request #180 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0-fix to release/2.7.0
  Merge pull request #178 in ADUC/open-apaas-extapp-ios from feature/2.7.0_toast to release/2.7.0
  Merge pull request #176 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0-fix to release/2.7.0
  Merge pull request #175 in ADUC/open-apaas-extapp-ios from feature/2.7.0_delete_file to release/2.7.0
  Merge pull request #177 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0_uiconfig to release/2.7.0
  Merge pull request #174 in ADUC/open-apaas-extapp-ios from hotfix/2.7.0_name to release/2.7.0
  ...

+++ AgoraWidgets/Cloud/AgoraCloudServerAPI.swift
@@ -101,7 +101,7 @@ extension AgoraCloudServerAPI {
     struct Conversion: Convertable {
         let type: String
         let preview: Bool
-        let scale: Double
+        let scale: Float
         let canvasVersion: Bool?
         let outputFormat: String
     }
